Move stream block generation code to another module
